+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "cats-searcher",
+  "name": "cat-searcher",
   "version": "1.0.0",
   "author": "Alexander Pop",
   "private": true,
