+++ public/index.html
@@ -5,14 +5,14 @@
     <!--    <link rel="icon" href="%PUBLIC_URL%/favicon.ico"/>-->
     <meta name="viewport" content="width=device-width, initial-scale=1"/>
     <meta
-            name="cats-searcher"
+            name="cat-searcher"
             content="web site for searching cats"
     />
     <!--    <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png"/>-->
 
     <!--    <link rel="manifest" href="%PUBLIC_URL%/manifest.json"/>-->
 
-    <title>Cats Searcher</title>
+    <title>Cat Searcher</title>
 </head>
 <body>
 <noscript>You need to enable JavaScript to run this app.</noscript>
