+++ src/components/Voting/Voting.module.scss
@@ -45,107 +45,4 @@
       border-radius: 20px;
     }
   }
-
-  &__nav_buttons {
-    position: absolute;
-    @include flex(flex-start, center);
-    bottom: -40px;
-    width: 248px;
-    height: 80px;
-    overflow: hidden;
-    border-radius: 20px;
-    border: 4px solid var(--bg_white-grey);
-    background: white;
-
-    .nav_button_wr {
-      @include flex(center, center);
-      width: 100%;
-      height: 100%;
-      margin-right: 4px;
-      cursor: pointer;
-      overflow: hidden;
-      position: relative;
-      outline: none;
-      border: none;
-
-      &:active {
-        svg {
-          transform: translateY(2px);
-        }
-      }
-
-      &:last-child {
-        margin-right: 0;
-      }
-
-      svg {
-        fill: white;
-      }
-
-      &.like_bg {
-        background: #97EAB9;
-      }
-
-      &.heart_bg {
-        background: #FF868E;
-
-        img {
-          padding: 15px;
-        }
-
-        &:active {
-          img {
-            transform: translateY(2px);
-          }
-        }
-      }
-
-      &.unlike_bg {
-        background: #FFD280;
-      }
-    }
-  }
-
-  &__message_wr {
-    background: var(--bg_color);
-    border-radius: 10px;
-    @include flex(flex-start, center);
-    padding: 15px 60px 15px 15px;
-    height: 60px;
-    position: relative;
-    margin-bottom: 10px;
-
-    &:last-child {
-      margin-bottom: 0;
-    }
-  }
-
-  &__time, &__desc {
-    display: inline-block;
-  }
-
-  &__time {
-    background: var(--bg_blackWhite);
-    padding: 3px 10px;
-    margin-right: 20px;
-    border-radius: 10px;
-    color: var(--title_color);
-  }
-
-  .id_desc {
-    color: var(--title_color);
-  }
-
-  .info__icon_wr {
-    content: '';
-    position: absolute;
-    width: 20px;
-    height: 20px;
-    top: 20px;
-    right: 20px;
-
-    svg {
-      fill: #97EAB9;
-    }
-  }
 }
