+++ src/components/Voting/VotingMessages/types.ts
@@ -1,0 +1,6 @@
+export type TVotingMessages = {
+	key: number,
+	id?: string,
+	message: string,
+	time: string,
+}
